fix: lsp config override in pyright
lsp-config was not using the config in the lsp/pyright.lua as it was being overridden by the default config. so i had to use either a different file name like lsp/pyright.lua or use vim.lsp.config(pyright, ...) to override the default not sure if it is a bug or an config error. i ended up using the latter